FyBlog/FyBlog/requirements.txt
amqp==2.5.1
anyjson==0.3.3
asn1crypto==0.24.0
attrs==19.1.0
Automat==0.7.0
billiard==3.6.1.0
celery==4.3.0
certifi==2019.9.11
cffi==1.12.3
chardet==3.0.4
Click==7.0
constantly==15.1.0
cryptography==2.7
cssselect==1.1.0
cycler==0.10.0
Django==2.1.1
django-celery==3.3.1
django-celery-results==1.1.2
django-contrib-comments==1.9.1
django-mdeditor==0.1.15
django-redis==4.10.0
dnspython==1.16.0
eventlet==0.25.0
greenlet==0.4.15
hyperlink==19.0.0
idna==2.8
importlib-metadata==0.19
incremental==17.5.0
kiwisolver==1.1.0
kombu==4.6.4
lxml==4.4.1
Markdown==3.1.1
matplotlib==3.1.1
monotonic==1.5
mysqlclient==1.4.2
Naked==0.1.31
numpy==1.17.2
parsel==1.5.2
Pillow==6.1.0
pyasn1==0.4.7
pyasn1-modules==0.2.6
pycparser==2.19
pycryptodome==3.9.0
pycryptodomex==3.8.2
PyDispatcher==2.0.5
Pygments==2.4.2
PyHamcrest==1.9.0
PyMySQL==0.9.3
pyOpenSSL==19.0.0
pyparsing==2.4.2
pyqt5-tools==5.13.0.1.5
python-dateutil==2.8.0
python-dotenv==0.10.3
pytz==2019.1
PyYAML==5.1.2
queuelib==1.5.0
redis==3.3.7
requests==2.22.0
Scrapy==1.7.3
service-identity==18.1.0
shellescape==3.4.1
six==1.12.0
treelib==1.5.5
Twisted==19.7.0
urllib3==1.25.6
vine==1.3.0
w3lib==1.21.0
xlrd==1.2.0
zipp==0.5.2
zope.interface==4.6.0
